+++ packages/terra-core-docs/src/terra-dev-site/test/paginator/ProgressivePaginatorWithFusionAnchorStyle.test.jsx
@@ -3,7 +3,7 @@ import ProgressivePaginator from 'terra-paginator/lib/ProgressivePaginator';
 
 const ProgressivePaginatorExample = () => (
   // eslint-disable-next-line no-console
-  <ProgressivePaginator onPageChange={i => console.log(i)} fusionAnchor selectedPage={1} totalCount={2234} itemCountPerPage={20} />
+  <ProgressivePaginator onPageChange={i => console.log(i)} anchorStyle selectedPage={1} totalCount={2234} itemCountPerPage={20} />
 );
 
 export default ProgressivePaginatorExample;